added Blt_ObjIsInteger

+++ src/bltCanvLabel.c
@@ -703,8 +703,7 @@ ScaleToFit(LabelItem *labelPtr, double xScale, double yScale)
     double newFontSize;
     Blt_Font font;
 
-    newFontSize = (int)(MIN(xScale, yScale) *
-                        Blt_Font_PointSize(labelPtr->baseFont));
+    newFontSize = MIN(xScale, yScale) * Blt_Font_PointSize(labelPtr->baseFont);
     labelPtr->flags |= DISPLAY_TEXT;
     /* Create a scaled font and replace the base font with it. */
     font = Blt_Font_Duplicate(labelPtr->tkwin, labelPtr->baseFont,
